+++ docxbox/docx/docx_archive_replace.h
@@ -42,7 +42,8 @@ class docx_archive_replace : public docx_archive {
   // New image file added, requires adding resp. relation
   bool added_image_file_ = false;
 
-  bool AddImageFileAndRelation(const std::string &image_markup_json);
+  std::string AddImageFileAndRelation(const std::string &image_markup_json);
+
   bool hasArgOfAdditionalImageFile() const;
 
   void InitDocxOutPathForReplaceText(